Confirm dialog and tenant mangement improvements

+++ src/ui/src/components/AmberGlobalAdmin.vue
@@ -2,6 +2,8 @@
 import {ref, onMounted} from "vue"
 import { AmberClient, type Tenant} from "amber-client"
 import { globalTenant } from "../../../shared/src"
+import { uiHelper} from "@/state"
+import { state } from "@/state";
 var props = defineProps<{amberClient: AmberClient}>();
 
 var allTenants = ref<Tenant[]>([]);
@@ -10,6 +12,12 @@ var newTenantId = ref("");
 var adminApi = props.amberClient.getGlobalAdminApi()!;
 var deleteTenant = async (tenantId :string)=>{
  
+    if (!(await uiHelper.confirmDialog("Are you sure you want to delete tenant " + tenantId + "?")))
+    {
+      console.log("User cancelled delete tenant");
+      return;
+    }
+
     var result = await adminApi.deleteTenant(tenantId);
     if (result.success)
     {
@@ -31,6 +39,14 @@ var refreshTenants = async ()=>{
 onMounted(async ()=>{
   await refreshTenants();
 });
+
+var createLinkToTenantApp = (tenantId:string) :string =>
+{
+  var targetUrl = state.uiConfig.loginTargetUrl || "/";
+  targetUrl = targetUrl.replace('{tenant}', tenantId);
+    
+  return targetUrl;
+};
 </script>
 <template>
   <v-container>
@@ -46,6 +62,8 @@ onMounted(async ()=>{
             <v-list-item-title>{{tenant.name}} [{{ tenant.id }}]</v-list-item-title>
             <v-list-item-action>
               <v-btn icon="mdi-delete-outline" @click="deleteTenant(tenant.id)" title="delete" v-if="tenant.id != globalTenant"></v-btn>
+              <v-btn icon="mdi-application-edit" title="manage" :href='"admin?tenant=" + tenant.id'></v-btn>
+              <v-btn icon="mdi-open-in-new" title="visit" :href='createLinkToTenantApp(tenant.id)'></v-btn>
             </v-list-item-action>
           </v-list-item>
         </v-list>
